memori-api/requirements.txt
memorisdk==2.1.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
mysql-connector-python==9.4.0
psycopg2-binary==2.9.10